Naming things: Rename package to `loko`
