+++ tests/test_util.py
@@ -3,8 +3,8 @@ import jq
 import jsonpointer
 import pytest
 import transon
-from zyp.util.expression import compile_expression
-from zyp.util.locator import to_pointer
+from loko.util.expression import compile_expression
+from loko.util.locator import to_pointer
 
 
 def test_to_pointer_string():
