+++ tests/transformation-bucket.json
@@ -1,7 +1,7 @@
 {
   "meta": {
     "version": 1,
-    "type": "zyp-bucket"
+    "type": "loko-bucket"
   },
   "schema": {
     "rules": [
